Change the logic of cell ID comparison
Now, we use cell IDs as the highest ranking comparison, and any cells that map equal will take priority in computing snakes of matching cells. Any further comparisons (strict, moderate, approximate) will be as before, but with the implication that the cells being compared do not have an ID match on the other side, so considering ID has no meaning.